+++ src/index.js
@@ -12,40 +12,8 @@ import { graphqlExpress, graphiqlExpress } from 'apollo-server-express';
 import { makeExecutableSchema } from 'graphql-tools';
 import resolvers from './graphql/resolvers';
 import typeDefs from './graphql/schema';
-import swagger from 'express-swagger-generator';
 
 let app = express();
-
-const expressSwagger = swagger(app);
-
-let options = {
-  swaggerDefinition: {
-    info: {
-      description: 'This is a sample server',
-      title: 'Swagger',
-      version: '1.0.0',
-    },
-    host: 'localhost:3000',
-    basePath: '/api',
-    produces: [
-      "application/json",
-      "application/xml"
-    ],
-    schemes: ['http', 'https'],
-    securityDefinitions: {
-      JWT: {
-        type: 'apiKey',
-        in: 'header',
-        name: 'Authorization',
-        description: "",
-      }
-    }
-  },
-  basedir: __dirname, //app absolute path
-  files: ['./api/**/*.js'] //Path to the API handle folder
-};
-expressSwagger(options);
-
 app.server = http.createServer(app);
 
 // logger
@@ -68,7 +36,7 @@ initializeDb( db => {
   app.use(middleware({ config, db }));
 
   // api router
-  app.use('/api', api({ config, db }));
+  app.use('/api', api({ config, db, app }));
   app.use('/img', img({ config, db }));
 
   const port = process.env.PORT || config.server.port
